+++ src/Resources/ResourceManager/SdkClient/DeploymentStacksSdkClient.cs
@@ -1030,7 +1030,7 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.SdkClient
             WriteVerbose("Starting DeploymentOperations polling");
             try
             {
-                ResourceManagerSdkClient.ProvisionDeploymentStatus(parameters, new Microsoft.Azure.Management.Resources.Models.Deployment());
+                ResourceManagerSdkClient.ProvisionDeploymentStatus(parameters, new Microsoft.Azure.Management.ResourceManager.Models.Deployment());
             }
             catch (Exception)
             {